Add game settings and improve collectible spawning logic
Introduced new game settings files, BurstAotSettings_WebGL.json and CommonBurstAotSettings.json, to manage Burst AOT and common settings. Also, updated the logic for spawning collectibles in the MainGame scene to make it more dynamic. The spawning now happens at regular intervals (that can be adjusted in the 'CollectibleSpawner' script) instead of fixed time frames. This is intended to add unpredictability to the gameplay and increase player engagement. Additionally, the movement speed of objects and screen bounds can now be manipulated for better game management.